cloud-brain/src/requirements.txt
aws-lambda-powertools==2.31.0
boto3>=1.28.0
pydantic>=2.0.0
brotli>=1.1.0
cryptography>=41.0.0
PyJWT>=2.8.0
